feat: add hook useSound
