+++ src/screens/donations/DonationDoneScreen/index.tsx
@@ -1,10 +1,10 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { RootStackScreenProps } from "types";
 import { useNavigation } from "hooks/useNavigation";
 import { useTranslation } from "react-i18next";
 import DoneScreenTemplate from "screens/templates/DoneScreenTemplate";
 import useFormattedImpactText from "hooks/useFormattedImpactText";
-import { Audio } from 'expo-av';
+import useSound from "hooks/useSound";
 
 export default function DonationDoneScreen({
   route,
@@ -12,26 +12,10 @@ export default function DonationDoneScreen({
   const { t } = useTranslation("translation", {
     keyPrefix: "donations.donationDoneScreen",
   });
-  const [sound, setSound] = useState<any>();
   const { nonProfit } = route.params;
   const { navigateTo } = useNavigation();
   const { formattedImpactText } = useFormattedImpactText();
-
-  const playSound = useCallback(async () => {
-    const { sound } = await Audio.Sound.createAsync(require('./assets/donation-done.mp3')
-    );
-    setSound(sound);
-
-    await sound.playAsync();
-  }, []);
-
-  useEffect(() => {
-    return sound
-      ? () => {
-        sound.unloadAsync();
-      }
-      : undefined;
-  }, [sound]);
+  const { playSound } = useSound();
 
   useEffect(() => {
     setTimeout(() => {
